astroid: 2.5.1 -> 2.5.3
This does not update to the latest version from the 2.5.x range,
since the build fails with that. It appears to be trying to access
an url, which is not allowed in the sandbox.

Since this is just a drive-by, I think this is better than keeping
all those packages broken on hydra, since the build is broken with 2.5.1