cbqn: build using clang in the native case
gcc generates a lot of noisy warnings which are speculative (*may* be
unintialized etc.). clang is much quieter here since it only generates a
warning if it can be sure, it seems.

In the cross case we happily use the default stdenv, since there often
is no binary cache for the clang stdenvs.