coreboot-toolchain: Update runtime dependencies
Add the build option `installExtraPackages`, which allows switching
between a "basic" and an "extended" toolchain. This option is set to
true by default adding additional packages to propagatedBuildInputs.
These packages are essential to make the toolchain fully functional.
Most developers and users will need them. For some usecases, these
packages might not be needed. So this option can be set to false.

This includes:

  * git
  * ncurses
  * pkg-config
  * python3Minimal

Signed-off-by: Felix Singer <felixsinger@posteo.net>