factorio: 1.1.39 -> 1.1.42
Also fixes a small issue in the update.py:
The upstream API seems to have changed in a sense that the demo is not
available in experimental channel anymore. Hence the original update.py
script would fail, due to the missing `demo` key in the experimental
channel. I added a sane failure path which just ignores if a release
type (like `demo`) is missing in the channel, in order to increase the
scripts resilience.