dbqn: jdk is always a native build input
We always need to execute javac on the build platform, so it has to be
native. The produced .jar should be platform independent, so it can
later be executed on the host platform.

Later, `${lib.getBin jdk}` will use the jdk from pkgsHostTarget (since
this is the default case in the splicing logic) which can be executed on
the host platform — exactly what we want.

Of course this is theoretical, since some dependencies of dbqn don't
cross compile in nixpkgs at the moment.