catgirl: fix path to openssl utility after configuring
catgirl wants to invoke the openssl utility at runtime and tries to
obtain a path to the binary as OPENSSL_BIN. This uses pkg-config's
exec_prefix which is not where binaries are installed in nixpkgs,
sadly. There is (at least as far as I know) no more appropriate
pkg-config variable unfortunately.