foot: 1.9.0 -> 1.9.2
foot's upstream has abandoned the approach of using a custom terminfo
install location and setting TERMINFO, but still supports it. For us
this has eliminated the need to build the foot terminfo files manually,
somehow which is appreciated.

We preserve our current approach: Install terminfo files to the terminfo
output and set TERMINFO to that store path instead of propagating the
output for user env insatllations.

Change logs:

* https://codeberg.org/dnkl/foot/releases/tag/1.9.1

* https://codeberg.org/dnkl/foot/releases/tag/1.9.2