Drop boehmgc_766 attribute
Reason:
- boehmgc_766 (pkgs/top-level/all-packages.nix) is only defined but not
used in Nixpkgs, so dropping it will not affect any package in Nixpkgs.
It was introduced for the asymptote package which has moved on to the
regular boehmgc attribute. There's a slim chance that some user depends
on it, but it would be unrealistic to expect us to maintain the
attribute indefinitely.

Details:
- Delete pkgs/development/libraries/boehm-gc/7.6.6.nix
- Delete pkgs/development/libraries/boehm-gc/riscv.patch