nixosTests.gnome: fix
GNOME Shell 41 made the Eval command we use for NixOS tests internal
so in order to be still able to use it, we need to run the shell
in an unsafe mode.

https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/1970
https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1983

Hopefully, the test driver will eventually get better so that
we will not need hacks like this forever.