ekam: init at 2021-09-18
This partially addresses #79204 and is based on work at
https://github.com/zenhack/ekam-nix

The capnproto source is required to build ekam.  Specifically,
the capnproto git repository is required, as ekam rule files
are not included in capnproto source tarballs.  So, this commit
modifies the capnproto nix expression to fetch source from GitHub
instead of using a release tarball.  This should not impact
the capnproto expression in any adverse way.

Co-authored-by: Ian Denhardt <ian@zenhack.net>