stdenv: define NIX_MAX_CORES
When `build-cores` is less than the number of CPUs, `make -l$NIX_BUILD_CORES`
keeps the remaining CPUs unutilized even when `max-jobs` is greater than 1.
This is good if you want to dedicate those CPUs to something other than Nix
builds, and it is bad if you want to balance `build-cores` with `max-jobs` to
maximize utilization of the system without overloading it. To achieve the latter
goal `make -l` should be given the optimal load average which is now available
in `$NIX_MAX_CORES`.

The implementation mirrors Ninja:
https://github.com/ninja-build/ninja/blob/e234a7bd/src/ninja.cc#L222-L233
https://github.com/ninja-build/ninja/blob/e234a7bd/src/util.cc#L473-L481

Co-authored-by: Orivej Desh <orivej@gmx.fr>