calibre: 5.24.0 -> 5.29.0
Also refresh the Debian patches. The last ones were from
calibre-{0.8.29,2.9.0}, and now they're up-to-date to version 5.29.0
(also downloading it directly from Debian bug tracker instead of copying
them).

As a side effect, plugins are supported again. I am assuming that if
Debian doesn't try to disable this anymore, they should be safer to use
now, or Debian decided that the risk is on the user. Either way, should
also be fine for us.