nixos/vmware-guest: fix headless option
The headless option broke with 7d8b303e3fd76ccf58cfe26348e889def3663546
because the path /bin/vmware-user-suid-wrapper does not exist in the
headless variant of the open-vm-tools package.

Since the vmblock fuse mount and vmware-user-suid-wrapper seem to only
be used for shared folders and drag and drop, they should not exist in
the vmware-guest module if it is configured as headless.