rakshasa-rtorrent: create subtree
RTorrent uses a companion library, libtorrent, and they should be synchronized
along the releases. They act like a small package set. Therefore it is a good
idea to treat them the same way in Nixpkgs code.

This commit should not change much of the things, because no code besides
rtorrent uses libtorrent.

(Yes, this the same message from jesec-rtorrent commit.)