lib: add `removeAll` as alias to `subtractLists`
I constantly find myself looking for a `remove` function that can take a
list of elements to remove, only to not find it and then remember that
`subtractLists` is what I'm looking for.

I figured we could introduce a `removeAll` as an alias to
`subtractLists`, this way users simply grepping for `remove` would hit
the correct function, and also usage when filtering `buildInputs` could
read more naturally.

Alternatively, we could just direct users to `subtractLists` in the
documentation for `remove`.

Not sure what our policy on aliased functions are, hopefuly this is
alright.