vscode-lldb: Switch to LLDB from LLVM 13
In 1.6.8 version of the extension, the author switched to LLVM 13. This
was necessary because the previously-released v1.6.7 version using LLVM
12 was broken.

See https://github.com/vadimcn/vscode-lldb/issues/527 for an example.

However, just upgrading the extension to 1.6.8 as nixpkgs has done is
not enough because we patch the LLDB being used to something that works
on NixOS. This _wasn't_ updated and still pointed at LLDB from LLVM 12.

The author is not updating the LLDB sources fork so I initially was
going to point at the official rustc fork though that'd require building
LLDB. Instead I just tried using the LLDB from nixpkgs directly and it
seems to be working. If someone spots some breakage, we should go back
to overriding the LLDB to point at the rustc fork. I'm not sure if they
are diverging at this moment.

I have confirmed locally that with this change, breakpoints work again
in Rust programs with the extension.