arcan.arcan: fix static openal issues
Arcan has some interesting idiosyncrasies. One of them is using cmake. The other
is using cmake to vendor some low-level, custom-patched dependencies.

In order to work around them, I have created ... a set of custom patches!
And on top of it, I have factored the vendored sources in a separate file, in
order to clean up the `default.nix` file.