nixos/systemd: Add systemd-userdbd.service
This provides the io.systemd.NameServiceSwitch Varlink service and
the io.systemd.Multiplexer varlink service.

This service is completely optional for now. Regardless of it being
enabled userdb records are being used internally by systemd. This just
exposes the API to third parties.

the service and nss_systemd can also read Systemd JSON user records drop
in files for /{run,etc}/userdb , which could be an interesting
alternative to our users perl script in the future. Though I don't want
to add that yet in this PR but maybe in a follow-up PR.

I thought of making the service optional. But the service is only
started through socket activation.