xmonad: Don't add ghc to the closure when 'config' is set
When 'services.xserver.windowManager.xmonad.config' is not null, the
restart feature won't work by default and is not desirable unless
specific care is made, as explained by the documentation.

I think it's reasonable to not include the haskell dependencies in the
environment by default. That reduces the size of my system by 2GB.