antimicrox: 3.0.1 -> 3.2.0
This changes the source from the (now archived) original antimicroX to a
fork that is actively maintained and includes new features like uinput
support for wayland.

Since upstream changed the executable name from antimicroX to
antimicrox, this also changes the name of the derivation and the file
names to work with Nix 2.4’s nix run and to be consistent. An alias for
antimicroX is added to ensure existing configurations will continue
working.