pkg-config: always pass --static in static builds
By default, pkg-config output will not include Requires.private
dependencies (which specify dependencies that only apply when building
statically).  It will only do this when passed --static.  In
pkgsStatic, let's ensure pkg-config is always operating in static
mode, because pkgsStatic will ensure that any libraries pkg-config
might find will always be static, and so will always need their
Requires.private dependencies.

This is very useful for Meson builds, because otherwise Meson will
only pass --static to pkg-config if the dependency was explicitly
"static : true", which is not likely to be the case for most stuff
we're building.

An alternative proposal was to patch Meson to add a special
environment variable to force dependencies to be "static : true".  I
feel that the approach I've taken here is less invasive.