vpnc: remove unnecessary patch
We can make the same changes much less invasively using normal Nixpkgs
stuff.  This would make it easier to switch to the actively maintained
fork[1] in future.

The use of $(out)/etc instead of /etc is a bit questionable, but I
don't want to change any behaviour in this patch so I'm leaving it
alone for now.

[1]: https://davidepucci.it/doc/vpnc/