linux: enable TCP_CONG_ADVANCED
TCP_CONG_ADVANCED is enabled by default on x86_64[1] in the upstream.
Although it is not the case for aarch64[2], many distributions, such as
Debian[3], Fedora[4] and Gentoo[5], choose to enable it in their
distribution kernel.

With this patch, aarch64 users can choose many other TCP congestion
algorithms, which may improve their network performance.

[1]: https://github.com/torvalds/linux/blob/7ddb58cb0ecae8e8b6181d736a87667cc9ab8389/arch/x86/configs/x86_64_defconfig#L71
[2]: https://github.com/torvalds/linux/blob/7ddb58cb0ecae8e8b6181d736a87667cc9ab8389/arch/arm64/configs/defconfig
[3]: https://salsa.debian.org/kernel-team/linux/-/blob/e2d14375d74cdcb2cfed6518d7a5ad784c30df1a/debian/config/config#L7063
[4]: https://src.fedoraproject.org/rpms/kernel/raw/836165dd2dff34e4f2c47ca8f9c803002c1e6530/f/kernel-aarch64-fedora.config
[5]: https://github.com/gentoo/gentoo/blob/5808eb2f066ea2ece19c35a1978b7f79077a467b/sys-kernel/gentoo-kernel/gentoo-kernel-5.10.77.ebuild#L27