find-tarballs.nix: Avoid all passthru attrs
Avoiding passthru itself was the right direction, but not a
complete solution. Passthru attributes typically do not contain
dependencies, but rather extra paths that are not relevant to
the build itself.

Whether we want to include all (passthru) test dependencies this
way can be debated, but removing them makes this expression more
robust.